Technology Transfer _ Sulambi VMS/Source Code/sulambi-backend-main/sulambi-backend-main/requirements.txt
flask
flask-cors
python-dotenv
scikit-learn
pandas
numpy
joblib
psycopg2-binary
gunicorn